# docs/aegis-read-only-bridge-openapi.yaml
openapi: 3.1.0
info:
  title: Aegis Read-Only ChatGPT Bridge
  version: "1.0.0"
  description: |
    Read-only local repository inspection bridge for Aegis Architect.
    The bridge does not provide write, execute, commit, push, shell, app launch,
    model-call, or external-network endpoints.
servers:
  - url: http://127.0.0.1:8765
security:
  - AegisBridgeToken: []
components:
  securitySchemes:
    AegisBridgeToken:
      type: apiKey
      in: header
      name: X-Aegis-Bridge-Token
      description: Local operator-provided bridge token. Do not place real tokens in docs.
  schemas:
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
    TreeNode:
      type: object
      properties:
        path:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [file, directory]
    SearchRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
          maxLength: 200
        globs:
          type: array
          items:
            type: string
          maxItems: 20
        max_results:
          type: integer
          minimum: 1
          maximum: 100
          default: 25
    SearchResult:
      type: object
      properties:
        path:
          type: string
        line_start:
          type: integer
        line_end:
          type: integer
        snippet:
          type: string
    ContextPackRequest:
      type: object
      properties:
        topic:
          type: string
          maxLength: 200
          nullable: true
        file_paths:
          type: array
          items:
            type: string
          maxItems: 20
        max_files:
          type: integer
          minimum: 1
          maximum: 25
          default: 8
        max_chars:
          type: integer
          minimum: 1
          maximum: 50000
          default: 12000
    ContextSnippet:
      type: object
      properties:
        path:
          type: string
        line_start:
          type: integer
        line_end:
          type: integer
        content:
          type: string
        truncated:
          type: boolean
paths:
  /health:
    get:
      operationId: getBridgeHealth
      summary: Get bridge health and deny policy metadata.
      description: Returns read-only bridge status, repository root, denied path policy, and current HEAD if available.
      responses:
        "200":
          description: Bridge health.
          content:
            application/json:
              example:
                status: ok
                bridge: aegis_read_only_chatgpt_bridge
                repo_root: C:\Users\nemes\Desktop\Aegis
                read_only: true
                runtime_dispatch_allowed: false
                execution_permission: not_granted_by_read_only_chatgpt_bridge
                write_endpoints: false
                command_execution_endpoint: false
                current_head: abc123
        "401":
          description: Missing or invalid token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /repo/status:
    get:
      operationId: getRepoStatus
      summary: Get read-only git status metadata.
      description: Returns branch, HEAD, clean/dirty state, and changed file paths. File contents are never returned.
      responses:
        "200":
          description: Repository status metadata.
          content:
            application/json:
              example:
                read_only: true
                branch: main
                head: abc123
                clean: false
                changed_files: [frontend/next-env.d.ts]
                ignored_runtime_artifacts_exposed_as_content: false
                secrets_exposed: false
  /repo/tree:
    get:
      operationId: getRepoTree
      summary: Get a filtered repository tree.
      parameters:
        - name: path
          in: query
          required: false
          schema:
            type: string
            default: "."
        - name: max_depth
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            maximum: 5
            default: 2
      responses:
        "200":
          description: Filtered tree response.
          content:
            application/json:
              schema:
                type: object
                properties:
                  read_only:
                    type: boolean
                  path:
                    type: string
                  nodes:
                    type: array
                    items:
                      $ref: "#/components/schemas/TreeNode"
                  truncated:
                    type: boolean
                  denylist_enforced:
                    type: boolean
  /repo/file:
    get:
      operationId: getRepoFile
      summary: Read a safe text file.
      description: Reads one allowlisted text file inside the repo. Denied paths, binary files, oversized files, and outside-repo paths are rejected.
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Safe text file content.
          content:
            application/json:
              example:
                read_only: true
                path: README.md
                size: 12067
                line_count: 300
                content: "# Aegis..."
                truncated: false
        "403":
          description: Denied path.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: File exceeds size limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "415":
          description: Binary or non-UTF-8 file rejected.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /repo/search:
    post:
      operationId: searchRepo
      summary: Search safe text files.
      description: Performs bounded Python text search over allowlisted text files. No shell query is constructed from user input.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequest"
      responses:
        "200":
          description: Search results.
          content:
            application/json:
              schema:
                type: object
                properties:
                  read_only:
                    type: boolean
                  query:
                    type: string
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/SearchResult"
                  result_count:
                    type: integer
                  denylist_enforced:
                    type: boolean
  /repo/context-pack:
    post:
      operationId: buildContextPack
      summary: Build a compact read-only context pack.
      description: Returns bounded snippets from explicit safe files or bounded search results. Context packs are not permission or authority.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContextPackRequest"
      responses:
        "200":
          description: Compact context pack.
          content:
            application/json:
              schema:
                type: object
                properties:
                  read_only:
                    type: boolean
                  context_package_is_permission:
                    type: boolean
                  snippets:
                    type: array
                    items:
                      $ref: "#/components/schemas/ContextSnippet"
                  total_chars:
                    type: integer
                  truncated:
                    type: boolean
  /repo/git-log:
    get:
      operationId: getRepoGitLog
      summary: Get recent commit metadata.
      description: Returns recent commit hashes, messages, and dates. It does not return diffs.
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        "200":
          description: Recent commit metadata.
          content:
            application/json:
              example:
                read_only: true
                limit: 10
                commits:
                  - hash: abc123
                    date: "2026-06-13T12:00:00+00:00"
                    message: Add read-only bridge
                diff_content_included: false